rpi/requirements.txt
pygame
pyserial
requests
